+++ tags/__init__.py
@@ -34,11 +34,13 @@ with open(Path(__file__).parent / "info.json") as fp:
     __red_end_user_data_statement__ = json.load(fp)["end_user_data_statement"]
 
 
-def setup(bot: Red) -> None:
+async def setup(bot: Red) -> None:
     cog = bot.get_cog("CustomCommands")
     if cog:
         raise CogLoadError(
             "This cog conflicts with CustomCommands and both cannot be loaded at the same time. "
             "After unloading `customcom`, you can migrate custom commands to tags with `[p]migratecustomcom`."
         )
-    bot.add_cog(Tags(bot))
+    tags = Tags(bot)
+    await tags.initialize()
+    bot.add_cog(tags)
